feat: remove deprecated GraphQLProject type
Complete the architecture refactor by removing the old GraphQLProject type
entirely from the codebase. All functionality has been migrated to either
StaticGraphQLProject (for CLI) or DynamicGraphQLProject (for LSP).

Changes:
- Remove GraphQLProject from lib.rs exports
- Delete project.rs file completely
- Update project_validation.rs tests to use StaticGraphQLProject
- Fix GraphQLConfig usage in tests (use Multi variant with HashMap)
- Add #[must_use] attributes to new methods
- Make getter methods const fn where appropriate
- Fix clippy warnings (needless borrows, type complexity)
- Add type aliases for complex LSP types
- Allow pedantic nursery lints for async RwLock patterns

All tests pass (208 tests). Clippy passes with no warnings.
