+++ crates/graphql-project/tests/fixtures/.graphqlrc.yaml
@@ -1,0 +1,2 @@
+schema: schema.graphql
+documents: "*.graphql"
